componetization and layout
