+++ src/main/resources/greboreda/agendafx/views/main.fxml
@@ -1,21 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.scene.control.Accordion?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
 <?import javafx.scene.control.Menu?>
 <?import javafx.scene.control.MenuBar?>
 <?import javafx.scene.control.MenuItem?>
 <?import javafx.scene.control.SeparatorMenuItem?>
-<?import javafx.scene.control.TableColumn?>
-<?import javafx.scene.control.TableView?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
 
-<?import javafx.scene.control.cell.PropertyValueFactory?>
-<VBox prefHeight="400.0" prefWidth="640.0" xmlns="http://javafx.com/javafx/9" xmlns:fx="http://javafx.com/fxml/1" fx:controller="greboreda.agendafx.controllers.MainController">
+<VBox alignment="TOP_CENTER" xmlns="http://javafx.com/javafx/9" xmlns:fx="http://javafx.com/fxml/1" fx:controller="greboreda.agendafx.controllers.MainController">
   <children>
     <MenuBar VBox.vgrow="NEVER">
       <menus>
@@ -56,37 +47,7 @@
         </Menu>
       </menus>
     </MenuBar>
-    <AnchorPane maxHeight="-1.0" maxWidth="-1.0" prefHeight="426.0" prefWidth="640.0" VBox.vgrow="ALWAYS">
-      <children>
-        <Label alignment="CENTER" layoutX="208.0" layoutY="31.0" style="&#10;" text="My awesome JavaFX app" textAlignment="CENTER" textFill="#9f9f9f" wrapText="false">
-          <font>
-            <Font size="18.0" />
-          </font>
-        </Label>
-            <Accordion layoutX="228.0" layoutY="186.0" />
-            <TableView fx:id="personsTable" layoutX="66.0" layoutY="86.0" prefHeight="200.0" prefWidth="523.0">
-              <columns>
-                <TableColumn prefWidth="75.0" text="id">
-                  <cellValueFactory>
-                    <PropertyValueFactory property="id" />
-                  </cellValueFactory>
-                </TableColumn>
-                <TableColumn prefWidth="196.0" text="First Name">
-                  <cellValueFactory>
-                    <PropertyValueFactory property="firstName" />
-                  </cellValueFactory>
-                </TableColumn>
-                <TableColumn prefWidth="251.0" text="Last Name">
-                  <cellValueFactory>
-                    <PropertyValueFactory property="lastName" />
-                  </cellValueFactory>
-                </TableColumn>
-              </columns>
-            </TableView>
-            <TextField fx:id="firstNameInput" layoutX="65.0" layoutY="310.0" prefHeight="26.0" prefWidth="203.0" promptText="First name" />
-            <TextField fx:id="lastNameInput" layoutX="288.0" layoutY="310.0" prefHeight="26.0" prefWidth="212.0" promptText="Last name" />
-            <Button fx:id="createPersonButton" layoutX="527.0" layoutY="310.0" mnemonicParsing="false" onAction="#onCreatePerson" text="Create" />
-      </children>
-    </AnchorPane>
+    <fx:include source="personsoutput.fxml" />
+    <fx:include source="personinput.fxml" />
   </children>
 </VBox>
