+++ src/main/resources/greboreda/agendafx/views/personsoutput.fxml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.control.cell.PropertyValueFactory?>
+<?import javafx.scene.layout.FlowPane?>
+
+
+<FlowPane alignment="TOP_CENTER" columnHalignment="CENTER" prefHeight="200.0" prefWidth="444.0" xmlns="http://javafx.com/javafx/9" xmlns:fx="http://javafx.com/fxml/1" fx:controller="greboreda.agendafx.controllers.PersonsOutputController">
+   <children>
+      <TableView fx:id="personsTable" prefHeight="200.0" prefWidth="523.0">
+          <columns>
+              <TableColumn maxWidth="75.0" prefWidth="75.0" text="id">
+                  <cellValueFactory>
+                      <PropertyValueFactory property="id" />
+                  </cellValueFactory>
+              </TableColumn>
+              <TableColumn maxWidth="1.7976931348623157E308" prefWidth="225.0" text="First Name">
+                  <cellValueFactory>
+                      <PropertyValueFactory property="firstName" />
+                  </cellValueFactory>
+              </TableColumn>
+              <TableColumn maxWidth="1.7976931348623157E308" prefWidth="225.0" text="Last Name">
+                  <cellValueFactory>
+                      <PropertyValueFactory property="lastName" />
+                  </cellValueFactory>
+              </TableColumn>
+          </columns>
+      </TableView>
+   </children>
+</FlowPane>
